+++ shuup_wishlist/plugins.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
-# This file is part of Shoop Wishlist.
+# This file is part of Shuup Wishlist.
 #
-# Copyright (c) 2012-2016, Shoop Ltd. All rights reserved.
+# Copyright (c) 2012-2016, Shoop Commerce Ltd. All rights reserved.
 #
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
@@ -10,32 +10,32 @@ from __future__ import unicode_literals
 from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
 
-from shoop.xtheme.resources import add_resource
-from shoop_wishlist.models import Wishlist
+from shuup.xtheme.resources import add_resource
+from shuup_wishlist.models import Wishlist
 
 try:
-    from shoop.xtheme import TemplatedPlugin
+    from shuup.xtheme import TemplatedPlugin
 except:
-    from shoop.xtheme.plugins import TemplatedPlugin
+    from shuup.xtheme.plugins import TemplatedPlugin
 
 
 def add_resources(context, content):
-    add_resource(context, "head_end", "%sshoop_wishlist/css/style.css" % settings.STATIC_URL)
-    add_resource(context, "body_end", "%sshoop_wishlist/js/lib.js" % settings.STATIC_URL)
-    add_resource(context, "body_end", "%sshoop_wishlist/js/flash_message.js" % settings.STATIC_URL)
+    add_resource(context, "head_end", "%sshuup_wishlist/css/style.css" % settings.STATIC_URL)
+    add_resource(context, "body_end", "%sshuup_wishlist/js/lib.js" % settings.STATIC_URL)
+    add_resource(context, "body_end", "%sshuup_wishlist/js/flash_message.js" % settings.STATIC_URL)
 
 
 class WishlistPlugin(TemplatedPlugin):
-    identifier = "shoop_wishlist.wishlist_button"
+    identifier = "shuup_wishlist.wishlist_button"
     name = _("Wishlist Plugin")
-    template_name = "shoop_wishlist/wishlist_action_button.jinja"
+    template_name = "shuup_wishlist/wishlist_action_button.jinja"
     required_context_variables = ['shop_product']
 
     def __init__(self, config):
         super(WishlistPlugin, self).__init__(config)
 
     def render(self, context):
-        add_resource(context, "body_end", "%sshoop_wishlist/js/script.js" % settings.STATIC_URL)
+        add_resource(context, "body_end", "%sshuup_wishlist/js/script.js" % settings.STATIC_URL)
         return super(WishlistPlugin, self).render(context)
 
     def get_context_data(self, context):