+++ shuup_wishlist_tests/settings.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
-# This file is part of Shoop Wishlist.
+# This file is part of Shuup Wishlist.
 #
-# Copyright (c) 2012-2016, Shoop Ltd. All rights reserved.
+# Copyright (c) 2012-2016, Shoop Commerce Ltd. All rights reserved.
 #
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
@@ -16,13 +16,13 @@ INSTALLED_APPS = (
     "django.contrib.contenttypes",
     "easy_thumbnails",
     "filer",
-    "shoop.core",
-    "shoop.front",
-    "shoop.customer_group_pricing",
-    "shoop.default_tax",
-    "shoop.simple_cms",
-    "shoop.xtheme",
-    "shoop_wishlist",
+    "shuup.core",
+    "shuup.front",
+    "shuup.customer_group_pricing",
+    "shuup.default_tax",
+    "shuup.simple_cms",
+    "shuup.xtheme",
+    "shuup_wishlist",
 )
 
 DATABASES = {
@@ -30,7 +30,7 @@ DATABASES = {
         'ENGINE': 'django.db.backends.sqlite3',
         'NAME': os.path.join(
             tempfile.gettempdir(),
-            'shoop_stripe_tests.sqlite3'
+            'shuup_stripe_tests.sqlite3'
         ),
     }
 }
@@ -49,7 +49,7 @@ MEDIA_ROOT = os.path.join(os.path.dirname(__file__), "var", "media")
 
 STATIC_URL = "static/"
 
-ROOT_URLCONF = 'shoop_workbench.urls'
+ROOT_URLCONF = 'shuup_workbench.urls'
 
 LANGUAGES = [
     ('en', 'English'),