+++ shuup_wishlist_tests/test_plugin.py
@@ -1,17 +1,17 @@
 # -*- coding: utf-8 -*-
-# This file is part of Shoop Wishlist.
+# This file is part of Shuup Wishlist.
 #
-# Copyright (c) 2012-2016, Shoop Ltd. All rights reserved.
+# Copyright (c) 2012-2016, Shoop Commerce Ltd. All rights reserved.
 #
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
 import pytest
 
-from shoop.core.models import get_person_contact
-from shoop.testing.factories import get_default_shop, get_default_shop_product
-from shoop.testing.utils import apply_request_middleware
-from shoop_wishlist.models import Wishlist, WishlistPrivacy
-from shoop_wishlist.plugins import WishlistPlugin
+from shuup.core.models import get_person_contact
+from shuup.testing.factories import get_default_shop, get_default_shop_product
+from shuup.testing.utils import apply_request_middleware
+from shuup_wishlist.models import Wishlist, WishlistPrivacy
+from shuup_wishlist.plugins import WishlistPlugin
 
 
 @pytest.mark.django_db