+++ CHANGELOG.rst
@@ -1,5 +1,11 @@
-Shoop Wishlist Change Log
+Shuup Wishlist Change Log
 =========================
+
+Version 0.2.0
+-------------
+
+Released on 2016-10-06
+Compatible with Shuup 0.5+
 
 Version 0.1.0
 -------------